Update Firstsolution.txt
This is an interesting problem, and there are several ways to approach it. Here is one potential solution to the problem:

Establish a rule that the first guest to put their cup in the first position gets the first cup of water, the second guest to put their cup in the second position gets the second cup of water, and so on.
Make sure that all the guests understand and agree to the rule.
Place a clear marking on the table to indicate the first, second, and last positions where the cups should be placed.
Ask the guests to place their cups on the table in a row, according to the established rule, and remind them that the cups should be placed in the designated positions, and that only the first and last cups will be filled.
Once all the cups are in place, pour water into the first and last cups, and let the guests drink from their cups.
This solution is based on the idea that the guests are rational and will try to maximize their chances of getting a cup of water. By establishing a clear rule and providing a clear indication of where the cups should be placed, guests are incentivized to act honestly and to race to put their cups in the designated positions.

In this situation, as you've mentioned, it's possible that some guests may try to cheat. But it's important to keep in mind that this solution is designed to be fair and maximizing the chances of getting a cup of water for everyone.

It's important to mention that if you are the host, you will have to act honestly and make sure that guests abide by the rules that you've set, while trying to provide a fair solution